# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TG-RSS Bot Monitoring API
  description: API для мониторинга и метрик Telegram RSS бота
  version: 1.0.0
  contact:
    name: TG-RSS Bot
    url: https://github.com/kaevdokimov/tg-rss

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /health:
    get:
      summary: Проверка здоровья сервиса
      description: Возвращает статус здоровья приложения и подключения к БД
      operationId: getHealth
      responses:
        "200":
          description: Сервис здоров
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
        "503":
          description: Сервис нездоров (проблемы с БД)
          content:
            text/plain:
              schema:
                type: string
                example: "Database unhealthy: connection refused"

  /metrics:
    get:
      summary: Получение метрик в формате Prometheus
      description: Возвращает метрики производительности в формате, совместимом с Prometheus
      operationId: getMetrics
      responses:
        "200":
          description: Метрики успешно получены
          content:
            text/plain:
              schema:
                type: string
                example: |
                  # TG-RSS Bot Metrics
                  # HELP rss_polls_total Total number of RSS polls
                  # TYPE rss_polls_total counter
                  rss_polls_total 150

                  # HELP rss_polls_errors_total Total number of RSS poll errors
                  # TYPE rss_polls_errors_total counter
                  rss_polls_errors_total 3

                  # HELP telegram_messages_sent_total Total number of Telegram messages sent
                  # TYPE telegram_messages_sent_total counter
                  telegram_messages_sent_total 1250

                  # HELP circuit_breaker_calls_total Total number of calls to circuit breaker rss_sources
                  # TYPE circuit_breaker_calls_total counter
                  circuit_breaker_calls_total{name="rss_sources"} 450

                  # HELP db_connections_open Current number of open database connections
                  # TYPE db_connections_open gauge
                  db_connections_open 5

  /api/v1/users:
    get:
      summary: Get all users
      parameters:
        - name: chat_id
          in: query
          schema:
            type: integer
          description: Filter by specific chat ID
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/sources:
    get:
      summary: Get all sources
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
    post:
      summary: Create new source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/sources/info:
    get:
      summary: Get source by ID
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Source information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/sources/update:
    put:
      summary: Update source
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceRequest'
      responses:
        "200":
          description: Source updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/sources/delete:
    delete:
      summary: Delete source
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Source deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/subscriptions:
    get:
      summary: Get user subscriptions
      parameters:
        - name: chat_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/subscriptions/subscribe:
    post:
      summary: Subscribe user to source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chat_id:
                  type: integer
                source_id:
                  type: integer
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/subscriptions/unsubscribe:
    delete:
      summary: Unsubscribe user from source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chat_id:
                  type: integer
                source_id:
                  type: integer
      responses:
        "200":
          description: Subscription deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

components:
  schemas:
    HealthResponse:
      type: string
      description: Статус здоровья сервиса
      enum:
        - "OK"
        - "Database unhealthy: <error_details>"

    MetricsResponse:
      type: string
      description: Метрики в формате Prometheus
      format: prometheus_metrics

    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string

    User:
      type: object
      properties:
        chat_id:
          type: integer
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        created_at:
          type: string
          format: date-time

    Source:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        news_count:
          type: integer
        subscribers:
          type: integer

    CreateSourceRequest:
      type: object
      required:
        - name
        - url
      properties:
        name:
          type: string
        url:
          type: string

    UpdateSourceRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        status:
          type: string

security: []